SW-26587 - Fix PHPStan issues with PHP 8 in components classes

+++ engine/Shopware/Components/Model/ModelManager.php
@@ -203,7 +203,7 @@ class ModelManager extends EntityManager
     {
         $metaDataCache = $this->getConfiguration()->getMetadataCacheImpl();
 
-        if (method_exists($metaDataCache, 'deleteAll')) {
+        if ($metaDataCache !== null && method_exists($metaDataCache, 'deleteAll')) {
             $metaDataCache->deleteAll();
         }
